update: indexAction 调用 index.js 中的init()
